+++ platform-operator/controllers/verrazzano/component/capi/capi_component.go
@@ -5,7 +5,6 @@ package capi
 
 import (
 	"context"
-	"encoding/base64"
 	"fmt"
 	vzconst "github.com/verrazzano/verrazzano/pkg/constants"
 	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
@@ -15,14 +14,10 @@ import (
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"os"
 	clusterapi "sigs.k8s.io/cluster-api/cmd/clusterctl/client"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/yaml"
 )
 
 // ComponentName is the name of the component
@@ -35,7 +30,6 @@ const ComponentNamespace = vzconst.CAPISystemNamespace
 const ComponentJSONName = "verrazzano-capi"
 
 const (
-	ociDefaultSecret                    = "oci"
 	capiCMDeployment                    = "capi-controller-manager"
 	capiKubeadmBootstrapCMDeployment    = "capi-kubeadm-bootstrap-controller-manager"
 	capiKubeadmControlPlaneCMDeployment = "capi-kubeadm-control-plane-controller-manager"
@@ -69,32 +63,6 @@ var capiDeployments = []types.NamespacedName{ // TODO: add OLCNE deployments
 		Name:      capiociCMDeployment,
 		Namespace: ComponentNamespace,
 	},
-}
-
-// AuthenticationType for auth
-type AuthenticationType string
-
-const (
-	// UserPrincipal is default auth type
-	UserPrincipal AuthenticationType = "user_principal"
-	// InstancePrincipal is used for instance principle auth type
-	InstancePrincipal AuthenticationType = "instance_principal"
-)
-
-// OCIAuthConfig holds connection parameters for the OCI API.
-type OCIAuthConfig struct {
-	Region      string             `yaml:"region"`
-	Tenancy     string             `yaml:"tenancy"`
-	User        string             `yaml:"user"`
-	Key         string             `yaml:"key"`
-	Fingerprint string             `yaml:"fingerprint"`
-	Passphrase  string             `yaml:"passphrase"`
-	AuthType    AuthenticationType `yaml:"authtype"`
-}
-
-// OCIConfig holds the configuration for OCI authorization.
-type OCIConfig struct {
-	Auth OCIAuthConfig `yaml:"auth"`
 }
 
 type capiComponent struct {
@@ -209,46 +177,7 @@ func (c capiComponent) IsInstalled(ctx spi.ComponentContext) (bool, error) {
 }
 
 func (c capiComponent) PreInstall(ctx spi.ComponentContext) error {
-	// Get OCI credentials from secret in the verrazzano-install namespace
-	ociSecret := corev1.Secret{}
-	// TODO: use secret name from API when available
-	if err := ctx.Client().Get(context.TODO(), client.ObjectKey{Name: ociDefaultSecret, Namespace: constants.VerrazzanoInstallNamespace}, &ociSecret); err != nil {
-		return ctx.Log().ErrorfNewErr("Failed to find secret %s in the %s namespace: %v", ociDefaultSecret, constants.VerrazzanoInstallNamespace, err)
-	}
-
-	var ociYaml []byte
-	for key := range ociSecret.Data {
-		if key == "oci.yaml" {
-			ociYaml = ociSecret.Data[key]
-			break
-		}
-	}
-
-	if ociYaml == nil {
-		return ctx.Log().ErrorfNewErr("Failed to find oci.yaml in secret %s in the %s namespace", ociDefaultSecret, constants.VerrazzanoInstallNamespace)
-	}
-
-	cfg := OCIConfig{}
-	if err := yaml.Unmarshal(ociYaml, &cfg); err != nil {
-		return ctx.Log().ErrorfNewErr("Failed to parse oci.yaml in secret %s in the %s namespace", ociDefaultSecret, constants.VerrazzanoInstallNamespace)
-	}
-
-	if cfg.Auth.AuthType == UserPrincipal {
-		os.Setenv("OCI_TENANCY_ID_B64", base64.StdEncoding.EncodeToString([]byte(cfg.Auth.Tenancy)))
-		os.Setenv("OCI_CREDENTIALS_FINGERPRINT_B64", base64.StdEncoding.EncodeToString([]byte(cfg.Auth.Fingerprint)))
-		os.Setenv("OCI_USER_ID_B64", base64.StdEncoding.EncodeToString([]byte(cfg.Auth.User)))
-		os.Setenv("OCI_REGION_B64", base64.StdEncoding.EncodeToString([]byte(cfg.Auth.Region)))
-		os.Setenv("OCI_CREDENTIALS_KEY_B64", base64.StdEncoding.EncodeToString([]byte(cfg.Auth.Key)))
-		if len(cfg.Auth.Passphrase) != 0 {
-			os.Setenv("OCI_PASSPHRASE_B64", base64.StdEncoding.EncodeToString([]byte(cfg.Auth.Passphrase)))
-		}
-	} else if cfg.Auth.AuthType == InstancePrincipal {
-		os.Setenv("USE_INSTANCE_PRINCIPAL_B64", base64.StdEncoding.EncodeToString([]byte("true")))
-	} else {
-		return ctx.Log().ErrorfNewErr("Invalid authtype value %s found for oci.yaml in secret %s in the %s namespace", cfg.Auth.AuthType, ociDefaultSecret, constants.VerrazzanoInstallNamespace)
-	}
-
-	return nil
+	return preInstall(ctx)
 }
 
 func (c capiComponent) Install(_ spi.ComponentContext) error {
